+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -169,6 +169,7 @@ export function withExtendedResolver(
     web: {
       'react-native': 'react-native-web',
       'react-native/index': 'react-native-web',
+      'react-native/Libraries/Image/resolveAssetSource': 'expo-asset/build/resolveAssetSource',
     },
   };